change test to advection field
