+++ test/test_helpers.py
@@ -297,40 +297,3 @@ def test_velocity_field_convert_input_error_when_callable_doesnt_return_fem_func
         match=f"A time dependent advection field should return an fem.Function, not a <class 'ufl.algebra.Product'>",
     ):
         test_value.convert_input_value(function_space=test_function_space, t=t)
-
-
-def test_velocity_field_update():
-    """Test a explicit time dependent input value is updated correctly"""
-
-    test_mesh = dolfinx.mesh.create_unit_square(MPI.COMM_WORLD, 10, 10)
-    V = fem.functionspace(test_mesh, ("Lagrange", 1))
-
-    v_cg = basix.ufl.element(
-        "Lagrange", test_mesh.topology.cell_name(), 2, shape=(test_mesh.geometry.dim,)
-    )
-    V_velocity = fem.functionspace(test_mesh, v_cg)
-    u = fem.Function(V_velocity)
-
-    def velocity_func_alt(x, t):
-        values = np.zeros((2, x.shape[1]))  # Initialize with zeros
-
-        scalar_value = 2 * t  # Compute the scalar function
-        values[0] = scalar_value  # Assign to first component
-        values[1] = scalar_value  # Second component remains zero
-
-        return values
-
-    def example_func(t):
-        u.interpolate(lambda x: velocity_func_alt(x, t))
-        return u
-
-    test_value = F.VelocityField(input_value=lambda t: example_func(t))
-    t = F.as_fenics_constant(value=1.0, mesh=test_mesh)
-
-    test_value.convert_input_value(function_space=V, t=t)
-
-    t_values = [1, 2, 3, 4, 5]
-
-    for t in t_values:
-        test_value.update(t)
-        assert np.isclose(np.max(test_value.fenics_object.x.array), 2 * t)
